Separate core, BC and Spring modules
Also set up module publications
