+++ BotCommands-typesafe-messages/bc/build.gradle.kts
@@ -1,0 +1,35 @@
+import org.jetbrains.kotlin.gradle.dsl.JvmTarget
+
+plugins {
+    id("BotCommands-conventions")
+    id("BotCommands-publish-conventions")
+}
+
+dependencies {
+    // -------------------- CORE DEPENDENCIES --------------------
+
+    api(projects.botCommandsCore)
+    api(projects.botCommandsTypesafeMessages.core)
+
+    // Logging
+    implementation(libs.kotlin.logging)
+
+    // -------------------- TEST DEPENDENCIES --------------------
+
+    testImplementation(libs.mockk)
+    testImplementation(libs.logback.classic)
+}
+
+java {
+    sourceCompatibility = JavaVersion.VERSION_24
+    targetCompatibility = JavaVersion.VERSION_24
+}
+
+kotlin {
+    compilerOptions {
+        jvmTarget = JvmTarget.JVM_24
+        optIn.add("dev.freya02.botcommands.typesafe.messages.api.annotations.ExperimentalTypesafeMessagesApi")
+    }
+}
+
+configurePublishedArtifact("BotCommands-typesafe-messages-bc")
