+++ BotCommands-typesafe-messages/bc/src/main/resources/META-INF/services/io.github.freya022.botcommands.api.core.reflect.ClassGraphProcessorProvider
