+++ BotCommands-typesafe-messages/bc/src/test/resources/logback-test.xml
