+++ BotCommands-typesafe-messages/core/build.gradle.kts
@@ -1,0 +1,34 @@
+import org.jetbrains.kotlin.gradle.dsl.JvmTarget
+
+plugins {
+    id("BotCommands-conventions")
+    id("BotCommands-publish-conventions")
+}
+
+dependencies {
+    // -------------------- CORE DEPENDENCIES --------------------
+
+    api(projects.botCommandsCore)
+
+    // Logging
+    implementation(libs.kotlin.logging)
+
+    // -------------------- TEST DEPENDENCIES --------------------
+
+    testImplementation(libs.mockk)
+    testImplementation(libs.logback.classic)
+}
+
+java {
+    sourceCompatibility = JavaVersion.VERSION_24
+    targetCompatibility = JavaVersion.VERSION_24
+}
+
+kotlin {
+    compilerOptions {
+        jvmTarget = JvmTarget.JVM_24
+        optIn.add("dev.freya02.botcommands.typesafe.messages.api.annotations.ExperimentalTypesafeMessagesApi")
+    }
+}
+
+configurePublishedArtifact("BotCommands-typesafe-messages-core")
