+++ BotCommands-typesafe-messages/core/src/main/kotlin/dev/freya02/botcommands/typesafe/messages/api/exceptions/AbstractMessageSourceMethodException.kt
