+++ BotCommands-typesafe-messages/core/src/test/kotlin/dev/freya02/botcommands/typesafe/messages/MessageSourceGeneratorTest.kt
