+++ BotCommands-typesafe-messages/spring/build.gradle.kts
@@ -1,0 +1,41 @@
+import org.jetbrains.kotlin.gradle.dsl.JvmTarget
+
+plugins {
+    id("BotCommands-conventions")
+    id("BotCommands-publish-conventions")
+}
+
+dependencies {
+    // -------------------- CORE DEPENDENCIES --------------------
+
+    api(projects.botCommandsCore)
+    api(projects.botCommandsTypesafeMessages.core)
+
+    // Logging
+    implementation(libs.kotlin.logging)
+
+    implementation(libs.spring.boot)
+    implementation(libs.spring.boot.autoconfigure)
+
+    // -------------------- TEST DEPENDENCIES --------------------
+
+    testImplementation(libs.mockk)
+    testImplementation(libs.logback.classic)
+
+    testImplementation(projects.botCommandsSpring)
+    testImplementation(libs.spring.boot.starter.test)
+}
+
+java {
+    sourceCompatibility = JavaVersion.VERSION_24
+    targetCompatibility = JavaVersion.VERSION_24
+}
+
+kotlin {
+    compilerOptions {
+        jvmTarget = JvmTarget.JVM_24
+        optIn.add("dev.freya02.botcommands.typesafe.messages.api.annotations.ExperimentalTypesafeMessagesApi")
+    }
+}
+
+configurePublishedArtifact("BotCommands-typesafe-messages-spring")
