+++ BotCommands-typesafe-messages/spring/src/main/kotlin/dev/freya02/botcommands/typesafe/messages/internal/autoconfigure/TypesafeMessagesAutoConfiguration.kt
@@ -1,0 +1,16 @@
+package dev.freya02.botcommands.typesafe.messages.internal.autoconfigure
+
+import org.springframework.boot.autoconfigure.AutoConfiguration
+import org.springframework.context.ApplicationContext
+import org.springframework.context.annotation.Bean
+
+@AutoConfiguration
+internal open class TypesafeMessagesAutoConfiguration {
+
+    @Bean
+    internal open fun springMessageSourceFactoryProvider(
+        context: ApplicationContext,
+    ): MessageSourceFactoryPostProcessor {
+        return MessageSourceFactoryPostProcessor(context)
+    }
+}
