+++ build.gradle.kts
@@ -31,6 +31,9 @@ dependencies {
     dokka(projects.botCommandsCore)
     dokka(projects.botCommandsSpring)
     dokka(projects.botCommandsJdaKtx)
+    dokka(projects.botCommandsTypesafeMessages.core)
+    dokka(projects.botCommandsTypesafeMessages.bc)
+    dokka(projects.botCommandsTypesafeMessages.spring)
 }
 
 tasks.withType<Test> {
